sdex_ui_test_pytest/requirements.txt
pytest==5.3.5
allure-pytest==2.8.9
selenium=3.141.0
